[tinystory] Finish first write of story parser
It's not clean, it's not the greatest, and I want to change the
numeric fragments to actual text labels, as I think those would be
more user friendly. The statement terminators right now are
semicolons, and that is a bad choice, because text in general may have
semicolons (as natural language tends to do so). Perhaps the `#`
symbol might be a better choice for the future.