[psy] Parallelize comparedirs
Implemented a simple worker pool that parallelizes the sha256summing of
incoming files.  Started off as a joke, as I was expecting IO to
bottleneck this, but I've seen actual improvements.  Very anecdotally
running this in parallel on a ~25Gig dir:

    real    1m17.671s
    user    2m49.779s
    sys     0m24.327s

versus the previous implementation:

    real    2m23.283s
    user    1m51.259s
    sys     0m14.703s

shows some gains.  60 seconds are 60 seconds.  It might be possible to
improve this more, but this is good enough for now.